feat: Implement token refresh logic in App component and remove unused tokenRefresh hook
