+++ app/src/features/authSlice.js
@@ -31,5 +31,5 @@ const authSlice = createSlice({
   },
 });
 
-export const { setUser, clearUser, setAuthLoading, setAuthError } = authSlice.actions;
+export const { setUser, clearUser, setAuthLoading, setAuthError, setAccess } = authSlice.actions;
 export default authSlice.reducer;
